+++ src/components/AddContent.vue
@@ -3,7 +3,7 @@
     <h4>Add {{ type }}</h4>
     <div class="row gutter-sm">
       <div class="col-12">
-        <q-input type="text" float-label="Title" v-model="title" dark />
+        <q-input type="text" float-label="Title" v-model="title" />
       </div>
       <div class="col-12">
         <q-btn color="primary" class="float-right" @click.native="add">Add {{ type }}</q-btn>
@@ -20,7 +20,8 @@ export default {
   // name: 'ComponentName',
   data () {
     return {
-      title: ''
+      title: '',
+      types: ['rseries', 'oseries', 'olesson', 'osermon', 'oscratch']
     }
   },
   mounted () {
@@ -31,80 +32,20 @@ export default {
       this.title = ''
     },
     add () {
-      switch (this.type) {
-        case 'rseries':
-          this.addRSeries()
-          break
-        case 'oseries':
-          this.addOSeries()
-          break
-        case 'olessons':
-          this.addOLesson()
-          break
-        case 'osermons':
-          this.addOSermon()
-          break
-        default:
-          console.error('Incorrect add type')
-      }
-    },
-    addRSeries () {
-      var obj = {
-        title: this.title
-      }
-      this.$database.add('rseries', obj, (res) => {
-        this.modalFin()
-        Notify.create({
-          message: 'REAL Series created!',
-          type: 'positive',
-          positon: 'bottom-left'
+      if (this.types.includes(this.type)) {
+        var obj = {
+          title: this.title
+        }
+        this.$database.add(this.type, obj, (res) => {
+          this.modalFin()
+          Notify.create({
+            message: this.type + ' created!',
+            type: 'positive',
+            position: 'bottom-left'
+          })
+          this.$router.push({ name: this.type, params: { id: res._id } })
         })
-        this.$router.push({ name: 'rseries', params: { id: res._id } })
-      })
-    },
-    addOSeries () {
-      var obj = {
-        title: this.title
       }
-      this.$database.add('oseries', obj, (res) => {
-        this.modalFin()
-        Notify.create({
-          message: 'Other Series created!',
-          type: 'positive',
-          positon: 'bottom-left'
-        })
-        this.$router.push({ name: 'oseries', params: { id: res._id } })
-      })
-    },
-    addOLesson () {
-      console.log('adding...')
-      var obj = {
-        title: this.title
-      }
-      this.$database.add('olesson', obj, (res) => {
-        this.modalFin()
-        Notify.create({
-          message: 'Other Lesson created!',
-          type: 'positive',
-          positon: 'bottom-left'
-        })
-        this.$router.push({ name: 'olesson', params: { id: res._id } })
-      })
-    },
-    addOSermon () {
-      console.log('adding...')
-      var obj = {
-        title: this.title
-      }
-      this.$database.add('osermon', obj, (res) => {
-        this.modalFin()
-        Notify.create({
-          message: 'Other Sermon created!',
-          type: 'positive',
-          positon: 'bottom-left'
-        })
-        this.$router.push({ name: 'osermon', params: { id: res._id } })
-      })
     }
   }
 }