+++ src/components/Editor.vue
@@ -1,6 +1,7 @@
 <template>
   <div>
     <vue-editor
+      ref="editor"
       :editorToolbar="toolbarContent"
       :value="text"
       @input="$emit('update:text', $event)"
@@ -16,7 +17,7 @@ export default {
     VueEditor
   },
   // name: 'ComponentName',
-  props: [ 'text' ],
+  props: [ 'text', 'save' ],
   data () {
     return {
       content: this.text,
@@ -26,7 +27,25 @@ export default {
         ['blockquote'],
         [{ 'list': 'ordered' }, { 'list': 'bullet' }],
         ['clean']
-      ]
+      ],
+      editCount: 0
+    }
+  },
+  mounted () {
+    console.log('editor', this.$refs.editor)
+    if (this.save) {
+      this.$refs.editor.quill.on('selection-change', (range, oldRange, source) => {
+        if (range === null && oldRange !== null) {
+          this.editCount = 0
+          this.save()
+        }
+      })
+      this.$refs.editor.quill.on('text-change', (delta, oldDelta, source) => {
+        if (this.editCount >= 100) {
+          this.editCount = 0
+          this.save()
+        } else this.editCount++
+      })
     }
   }
 }