+++ src/components/modules/Quote.vue
@@ -1,5 +1,5 @@
 <template>
-  <q-card color="dark">
+  <q-card>
     <div v-show="!data.editing || data.editing !== $firebase.auth.currentUser.uid">
       <q-card-title>
         <q-icon v-show="!data.editing" class="float-right cursor-pointer" name="fas fa-ellipsis-v" color="primary" size="1rem">
@@ -24,10 +24,10 @@
         <q-icon link class="float-right cursor-pointer" name="fas fa-times" color="primary" size="1rem" @click.native="close" />
         <div class="row gutter-sm">
           <div class="col-12">
-            <q-input v-model="media.text" float-label="Quote" type="textarea" :max-height="100" :min-rows="1" dark />
+            <q-input v-model="media.text" float-label="Quote" type="textarea" :max-height="100" :min-rows="1" />
           </div>
           <div class="col-12">
-            <q-input v-model="media.notes" float-label="Notes" type="textarea" :max-height="100" :min-rows="1" dark />
+            <q-input v-model="media.notes" float-label="Notes" type="textarea" :max-height="100" :min-rows="1" />
           </div>
           <div class="col-12">
             <q-btn color="primary" @click.native="preSave">Save</q-btn>