+++ src/layouts/builder.vue
@@ -46,56 +46,60 @@
     <q-layout-drawer
       side="left"
       v-model="leftDrawer"
-      content-class="bg-grey-9 sm-drawer"
+      content-class="sm-drawer"
     >
       <!-- QScrollArea is optional -->
       <q-scroll-area class="fit q-pa-sm">
-        <q-item to="/dashboard">
+        <q-item to="/dashboard" class="menu-item">
           <q-item-tile icon="fa-home" />
           <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Dashboard</q-tooltip>
         </q-item>
-        <q-item to="/realdashboard" v-if="$root.user.realUser">
+        <q-item to="/real/dashboard" v-if="$root.user.realUser" class="menu-item">
           <q-item-tile image><img src="statics/logo_square@0.5x.png" style="width: 30px;" /></q-item-tile>
           <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">REAL Curriculum</q-tooltip>
         </q-item>
         <q-item-separator />
-        <q-item :to="{ name: 'olist', params: { type: 'sermons' }}">
+        <q-item :to="{ name: 'olist', params: { type: 'sermon' }}" class="menu-item">
           <q-item-tile icon="fa-comment" />
           <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Sermons</q-tooltip>
         </q-item>
-        <q-item :to="{ name: 'olist', params: { type: 'lessons' }}">
+        <q-item :to="{ name: 'olist', params: { type: 'lesson' }}" class="menu-item">
           <q-item-tile icon="fa-file" />
           <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Lessons</q-tooltip>
         </q-item>
+        <q-item :to="{ name: 'olist', params: { type: 'scratch' }}" class="menu-item">
+          <q-item-tile icon="fa-lightbulb" />
+          <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Scratch Pad</q-tooltip>
+        </q-item>
         <q-item-separator />
         <div v-if="!$root.user.nqUser">
-          <q-item :to="{ name: 'olist', params: { type: 'quotes' }}">
+          <q-item :to="{ name: 'olist', params: { type: 'quote' }}" class="menu-item">
             <q-item-tile icon="fa-quote-left" />
             <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Quotes</q-tooltip>
           </q-item>
-          <q-item :to="{ name: 'olist', params: { type: 'images' }}">
+          <q-item :to="{ name: 'olist', params: { type: 'image' }}" class="menu-item">
             <q-item-tile icon="fa-image" />
             <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Images</q-tooltip>
           </q-item>
-          <q-item :to="{ name: 'olist', params: { type: 'illustrations' }}">
+          <q-item :to="{ name: 'olist', params: { type: 'illustration' }}" class="menu-item">
             <q-item-tile icon="fa-paint-brush" />
             <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Illustrations</q-tooltip>
           </q-item>
-          <q-item :to="{ name: 'olist', params: { type: 'lyrics' }}">
+          <q-item :to="{ name: 'olist', params: { type: 'lyric' }}" class="menu-item">
             <q-item-tile icon="fa-music" />
             <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Lyrics</q-tooltip>
           </q-item>
-          <q-item :to="{ name: 'olist', params: { type: 'videos' }}">
+          <q-item :to="{ name: 'olist', params: { type: 'video' }}" class="menu-item">
             <q-item-tile icon="fa-play" />
             <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Videos</q-tooltip>
           </q-item>
         </div>
         <q-item-separator v-if="!$root.user.nqUser" />
-        <q-item>
+        <q-item to="/settings" class="menu-item">
           <q-item-tile icon="fa-cog" />
           <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Settings</q-tooltip>
         </q-item>
-        <q-item link @click.native="logout">
+        <q-item link @click.native="logout" class="menu-item">
           <q-item-tile icon="fa-sign-out-alt fab" />
           <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Log Out</q-tooltip>
         </q-item>
@@ -105,7 +109,7 @@
     <q-layout-drawer
       side="right"
       v-model="rightDrawer"
-      content-class="bg-grey-9 lg-drawer"
+      content-class="lg-drawer"
       :breakpoint="1200"
     >
       <q-scroll-area class="fit q-pa-sm">
@@ -144,6 +148,9 @@ export default {
       this.leftDrawer = val !== 'olesson'
       this.rightDrawer = val === 'olesson' && this.$q.platform.is.desktop && this.$root.user.nqUser
       this.showRightDrawer = val === 'olesson' && this.$q.platform.is.desktop
+    },
+    '$root.user.nqUser': function (val) {
+      this.rightDrawer = this.$route.name === 'olesson' && this.$q.platform.is.desktop && val
     }
   },
   methods: {
@@ -157,20 +164,16 @@ export default {
 }
 </script>
 
-<style>
+<style lang="stylus">
 
-.sm-drawer {
-  width: 75px
+@import "~variables"
+
+.menu-item:hover, .router-link-active {
+  background: inherit !important
 }
 
-.lg-drawer {
-  width: 100%;
-}
-
-@media screen and (min-width: 1200px) {
-  .lg-drawer {
-    width: 35%;
-  }
+.menu-item:hover > i, .router-link-active > i {
+  color: $primary !important
 }
 
 </style>