+++ src/pages/OList.vue
@@ -4,7 +4,7 @@
     <div v-if="loading">
       <q-spinner color="primary" class="absolute-center" size="3rem" />
     </div>
-    <div v-if="!loading">
+    <div v-if="!loading && contentTypes.includes(type)">
       <q-card inline v-for="item in items" :key="item._id" class="media-card" @click.native="openItem(item._id)">
         <q-card-title>{{ item.title }}</q-card-title>
         <q-card-main>
@@ -12,11 +12,23 @@
         </q-card-main>
       </q-card>
     </div>
+    <div v-if="!loading && mediaTypes.includes(type)">
+      <q-card inline v-for="item in items" :key="item._id" class="media-card" @click.native="openItem(item._id)">
+        <q-card-media v-if="type === 'image' || type === 'video'">
+          <img :src="item.thumbURL" />
+          <q-card-title slot="overlay" v-if="type === 'video'">{{ item.title }}</q-card-title>
+        </q-card-media>
+        <q-card-title v-if="type === 'lyric' || type === 'illustration'">{{ item.title }}</q-card-title>
+        <q-card-main v-if="type !== 'image' && type !== 'video'">
+          <p>{{ item.text }}</p>
+        </q-card-main>
+      </q-card>
+    </div>
     <q-modal ref="addContentModal" v-model="showAddContent" content-classes="add-media-modal">
-      <add-content :modal-fin="closeAdd" :type="'o' + type" ref="addContent" />
+      <add-content :modal-fin="closeAddContent" :type="'o' + type" ref="addContent" />
     </q-modal>
     <q-modal ref="addMediaModal" v-model="showAddMedia" content-classes="add-media-modal">
-      <!-- TODO: Add add-media component to manage adding new medias -->
+      <add-media :modal-fin="closeAddMedia" :type="type" ref="addMedia" />
     </q-modal>
     <q-modal ref="showMediaModal" v-model="showMedia" content-classes="add-media-modal">
       <!-- TODO: Add variable components to show and edit media -->
@@ -26,16 +38,18 @@
 
 <script>
 import AddContent from 'components/AddContent.vue'
+import AddMedia from 'components/AddMedia.vue'
 
 export default {
   components: {
-    AddContent
+    AddContent,
+    AddMedia
   },
   // name: 'PageName',
   data () {
     return {
-      contentTypes: ['rseries', 'oseries', 'olessons', 'osermons'],
-      mediaTypes: ['quotes', 'images', 'illustrations', 'lyrics', 'videos'],
+      contentTypes: ['series', 'lesson', 'sermon', 'scratch'],
+      mediaTypes: ['quote', 'image', 'illustration', 'lyric', 'video'],
       type: this.$route.params.type,
       items: [],
       loading: false,
@@ -46,6 +60,12 @@ export default {
   },
   mounted () {
     this.init(this.type)
+  },
+  watch: {
+    '$route.params.type' (newType, oldType) {
+      this.type = newType
+      this.init(newType)
+    }
   },
   methods: {
     init (type) {
@@ -59,18 +79,10 @@ export default {
     openItem (id) {
       console.log(id)
       console.log(this.type)
-      switch (this.type) {
-        case 'series':
-          this.$router.push({ name: 'oseries', params: { id: id } })
-          break
-        case 'lessons':
-          this.$router.push({ name: 'olesson', params: { id: id } })
-          break
-        case 'sermons':
-          this.$router.push({ name: 'osermon', params: { id: id } })
-          break
-        default:
-          console.error('Incorrect item type for routing')
+      if (this.contentTypes.includes(this.type)) {
+        this.$router.push({ name: 'o' + this.type, params: { id: id } })
+      } else if (this.mediaTypes.includes(this.type)) {
+        console.log('open media')
       }
     },
     openAdd () {
@@ -82,9 +94,12 @@ export default {
         console.error('incorrect type to add')
       }
     },
-    closeAdd () {
+    closeAddContent () {
       this.showAddContent = false
-      this.showAddContent = false
+    },
+    closeAddMedia (newItem) {
+      this.showAddMedia = false
+      this.items.push(newItem)
     }
   }
 }